backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
groq
psycopg2-binary
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose
supabase
stripe
gunicorn
